Directories fix
