+++ stream-chat-android-offline/build.gradle
@@ -122,5 +122,5 @@ tasks.register('dokkaDokkasaurus', DokkaTask) {
         dokkaPlugin(Dependencies.dokkasaurus)
     }
 
-    outputDirectory = file("../build/dokka/docusaurus")
+    outputDirectory = file("../build/dokka/docusaurus/stream-chat-android-offline")
 }
