+++ stream-chat-android-ui-components/build.gradle
@@ -104,5 +104,5 @@ tasks.register('dokkaDokkasaurus', DokkaTask) {
         dokkaPlugin(Dependencies.dokkasaurus)
     }
 
-    outputDirectory = file("../build/dokka/docusaurus")
+    outputDirectory = file("../build/dokka/docusaurus/stream-chat-android-ui-components")
 }
